+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## Unreleased
+
+### Fixed
+
+- Error with `Queue::createPayload` method
+- Error with incompatibility `RabbitQueueFailedJobProvider` with interface `Illuminate\Queue\Failed\FailedJobProviderInterface`
+
 ## v2.10.0
 
 ### Added